requirements.txt
ansible==9.4.0
ansible-core==2.16.6
ansible-runner==2.3.6
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
docutils==0.21.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Jinja2==3.1.3
lockfile==0.12.2
MarkupSafe==2.1.5
packaging==24.0
paramiko==3.4.0
pexpect==4.9.0
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyspnego==0.10.2
python-daemon==3.0.1
pywinrm==0.4.3
PyYAML==6.0.1
requests==2.31.0
requests-ntlm==1.2.0
resolvelib==1.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
xmltodict==0.13.0
zope.interface==6.3
